backend/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "realestate-backend"
version = "0.1.0"
description = "Real estate aggregator backend"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0,<6.0",
    "django-ninja>=1.3",
    "django-environ>=0.11",
    "django-extensions>=3.2",
    "django-celery-beat>=2.6",
    "django-ratelimit>=4.1",
    "psycopg[binary]>=3.2",
    "celery[redis]>=5.4",
    "redis>=5.0",
    "requests>=2.32",
    "pillow>=10.4",
    "python-slugify>=8.0",
    "anthropic>=0.40",
    "gunicorn>=22.0",
    "sentry-sdk[django]>=2.0",
    "owslib>=0.32",
]

[project.optional-dependencies]
gov = [
    "owslib>=0.32",
    "fiona>=1.10",
    "geopandas>=1.0",
    "pymupdf>=1.24",
    "pytesseract>=0.3",
    "anthropic>=0.40",
    "rapidfuzz>=3.10",
]
dev = [
    "ipython>=8.27",
    "ruff>=0.6",
    "pytest>=8.3",
    "pytest-django>=4.9",
    "model-bakery>=1.19",
]

[tool.setuptools.packages.find]
include = ["config*", "apps*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["*/migrations/*"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "DJ"]
